倍率支持小数

+++ gms-server/src/main/java/org/gms/model/dto/WorldListRtnDTO.java
@@ -11,11 +11,11 @@ import lombok.NoArgsConstructor;
 @Builder
 public class WorldListRtnDTO {
     private Integer id;
-    private Integer expRate;
-    private Integer dropRate;
-    private Integer mesoRate;
-    private Integer bossDropRate;
-    private Integer questRate;
-    private Integer travelRate;
-    private Integer fishingRate;
+    private Float expRate;
+    private Float dropRate;
+    private Float mesoRate;
+    private Float bossDropRate;
+    private Float questRate;
+    private Float travelRate;
+    private Float fishingRate;
 }
